refactor: Extract StockListItem into its own file
This commit extracts the `StockListItem` composable from `presentation/stock/StockScreen.kt` into its own file, `presentation/stock/StockListItem.kt`.

This refactoring improves code organization by separating the list item view into a dedicated component. A `@Preview` for the `StockListItem` has also been added in the new file for easier isolated development and testing.
